+++ lib/dartson.dart
@@ -5,6 +5,8 @@ import 'package:logging/logging.dart';
 export 'src/annotations.dart';
 import 'src/annotations.dart';
 import './type_transformer.dart';
+import 'src/reference_mapper.dart';
+import 'src/shared_exceptions.dart';
 
 @MirrorsUsed(metaTargets: const [Property], override: '*')
 import 'dart:mirrors';
@@ -20,6 +22,7 @@ class Dartson<T> {
   final Codec _codec;
   final Logger _log;
   final Map<String, TypeTransformer> _transformers = {};
+  final Map<String, Type> _types = {};
 
   Dartson(this._codec, [String identifier = 'dartson'])
       : _log = new Logger(identifier) {
@@ -40,87 +43,149 @@ class Dartson<T> {
   bool hasTransformer(Type type) =>
       _transformers[_getName(reflectType(type).qualifiedName)] != null;
 
+  /// Registers an identifier for the specific [type] to this darston instance.
+  /// Encoded result [T] will include these type information, and reused in [encode]/[decode] and [encodeReferenceAware]/[decodeReferenceAware].
+  addIdentifier(String identifier, Type type) {
+    _types[identifier] = type;
+  }
+
   /// Maps the values within [data] onto the [object] by reflecting the
   /// the type of object. The Class of [object] should have the [Entity]
   /// annotation to work properly when compiling to JavaScript.
+  /// Pass null for [object] will force using identifiers (see [addIdentifier] to create [object] instance.
   /// Returns the [object].
-  Object map(Object data, Object object, [bool isList = false]) {
-    var reflectee = reflect(object);
+  Object map(Object data, Object object, [bool isList = false, bool beReferenceAware = false]) {
+    var rootObject = object;
+    if (rootObject == null) {
+      var identifier = data is Map ? data[TYPE_KEY] : null;
+      if (identifier == null) {
+        throw new NullObjectError();
+      }
+      var type = _types[identifier];
+      if (type == null) {
+        throw new UnknownIdentifierError(identifier);
+      }
+      var c = reflectClass(type);
+      rootObject = c.newInstance(new Symbol(''), []).reflectee;
+    }
+    var reflectee = reflect(rootObject);
+    var mapper = beReferenceAware ? new DecodingReferenceMapper() : null;
 
     if (data is List && isList) {
       ClassMirror itemMirror = reflectee.type;
 
       var list = [];
       data.forEach((item) {
-        list.add(_convertValue(itemMirror, item, "@LIST_ITEM"));
+        list.add(_convertValue(itemMirror, item, "@LIST_ITEM", mapper));
       });
 
       return list;
     } else if (data is List && !isList) {
-      throw new IncorrectTypeTransform(object, 'List');
+      throw new IncorrectTypeTransform(rootObject, 'List');
     } else {
-      _fillObject(reflectee, data);
+      _fillObject(reflectee, data, mapper);
     }
 
-    return object;
+    return rootObject;
   }
 
   /// Transforms an [object] to a serializable map which can be handled
   /// by any encoder like JSON.
-  Object serialize(Object object) {
+  Object serialize(Object object, [EncodingReferenceMapper mapper]) {
     if (object is List) {
-      return _serializeList(object);
+      return _serializeList(object, mapper);
     } else if (object is Map) {
-      return _serializeMap(object);
+      return _serializeMap(object, mapper);
     } else if (object == null) {
       return null;
     } else {
-      return _serializeObject(object);
+      return _serializeObject(object, mapper);
     }
   }
 
-  /// Decodes the [endoded] object (for example a JSON encoded string) using
+  /// Decodes the [encoded] object (for example a JSON encoded string) using
   /// the [_codec] and then uses [map] to map it onto the [object].
-  Object decode(T encoded, Object object, [bool isList = false]) {
-    return map(_codec.decode(encoded), object, isList);
+  /// Reconstruct references and resolved placeholders added by [encodeReferenceAware].
+  Object decodeReferenceAware(T encoded, Object object, [ bool isList = false]) {
+    return decode(encoded, object, isList, true);
+  }
+
+  /// Decodes the [encoded] object (for example a JSON encoded string) using
+  /// the [_codec] and then uses [map] to map it onto the [object].
+  /// Use [addIdentifier] to register type.
+  /// Pass null for [object] will force using identifiers (see [addIdentifier] to create [object] instance.
+  Object decode(T encoded, Object object, [ bool isList = false, bool beReferenceAware =  false]) {
+    return map(_codec.decode(encoded), object, isList, beReferenceAware);
   }
 
   /// Serializes the [decoded] object using [serialize] and then calls the encode
   /// method on the [_codec].
-  T encode(Object decoded) {
-    return _codec.encode(serialize(decoded));
+  /// References to already mapped objects will be serialized by an placeholder (except Lists and Maps).
+  /// Optional use [addIdentifier] to register type.
+  T encodeReferenceAware(Object decoded) {
+    return encode(decoded, beReferenceAware: true);
+  }
+  /// Serializes the [decoded] object using [serialize] and then calls the encode
+  /// method on the [_codec].
+  /// Optional use [addIdentifier] to register type.
+  T encode(Object decoded, {bool beReferenceAware: false}) {
+    return _codec.encode(serialize(decoded, beReferenceAware ? new EncodingReferenceMapper() : null));
   }
 
   /// Creates a new list containing only serializable objects.
-  List _serializeList(List object) {
-    return object.map((i) => serialize(i)).toList();
+  List _serializeList(List object, EncodingReferenceMapper mapper) {
+    return object.map((i) => serialize(i, mapper)).toList();
   }
 
   /// Creates a new map containing only serializable objects.
-  Map _serializeMap(Map object) {
+  Map _serializeMap(Map object, EncodingReferenceMapper mapper) {
     var map = {};
     object.forEach((k, v) {
-      if (v != null) map[k] = serialize(v);
+      if (v != null) map[k] = serialize(v, mapper);
     });
     return map;
   }
 
-  Object _serializeObject(Object object) {
+  Object _serializeObject(Object object, EncodingReferenceMapper mapper) {
+    if (mapper != null && mapper.isSerialized(object)) {
+        return mapper.createSerializablePlaceholder(object);
+    }
     var reflectee = reflect(object);
     var symbolName = _getName(reflectee.type.qualifiedName),
         transformer;
 
     if ((transformer = _transformers[symbolName]) != null) {
-      return transformer.encode(object);
+      var result = transformer.encode(object);
+      if (mapper != null) {
+        mapper.registerSerializableMap(object, result);
+      }
+      return result;
     } else {
       Map result = new Map<String, Object>();
-      reflectee.type.declarations.forEach((sym, decl) {
-        if (!decl.isPrivate &&
-            ((decl is VariableMirror && !decl.isConst && !decl.isStatic) ||
-                (decl is MethodMirror && decl.isGetter))) {
-          _setField(sym, decl, reflectee, result);
-        }
-      });
+      if (mapper != null) {
+        mapper.registerSerializableMap(object, result);
+      }
+      var key = _types.keys.firstWhere((key) => _types[key] == object.runtimeType, orElse: () => null);
+      if (key != null) {
+        result.putIfAbsent(TYPE_KEY, () => key);
+      }
+
+      // serialize bottom to top in class hierarchy as dart will also by default
+      // initialize all field before constructor bodies are run i.e. this.fields before super.
+      // see also https://www.dartlang.org/dart-tips/dart-tips-ep-11.html
+
+      var objectClassMirror = reflectClass(Object);
+      var type = reflectee.type;
+      while (type != null) {
+        type.declarations.forEach((sym, decl) {
+          if (!decl.isPrivate &&
+              ((decl is VariableMirror && !decl.isConst && !decl.isStatic) ||
+                  (decl is MethodMirror && decl.isGetter))) {
+            _setField(sym, decl, reflectee, result, mapper);
+          }
+        });
+        type = type.superclass == objectClassMirror ? null : type.superclass;
+      }
 
       _log.finer("Serialization completed.");
       return result;
@@ -131,7 +196,7 @@ class Dartson<T> {
   /// the value to the [result] map. If there's no [Property] annotation
   /// with a different name set it will use the name of [symbol].
   void _setField(Symbol symbol, DeclarationMirror variable,
-      InstanceMirror instMirror, Map<String, Object> result) {
+      InstanceMirror instMirror, Map<String, Object> result, EncodingReferenceMapper mapper) {
     InstanceMirror field = instMirror.getField(symbol);
     Object value = field.reflectee;
     String fieldName = MirrorSystem.getName(symbol);
@@ -148,48 +213,61 @@ class Dartson<T> {
 
     if (value != null && (prop != null ? !prop.ignore : true)) {
       _log.finer("Serializing field: ${fieldName}");
-      result[fieldName] = serialize(value);
+      result[fieldName] = serialize(value, mapper);
     }
   }
 
   /// Puts the data of the [filler] into the object in [objMirror]
   /// Throws [IncorrectTypeTransform] if json data types doesn't match.
-  void _fillObject(InstanceMirror objMirror, Map filler) {
+  void _fillObject(InstanceMirror objMirror, Map filler, DecodingReferenceMapper mapper) {
+
+    if (mapper != null) {
+       mapper.registerInstanceIfApplicable(objMirror.reflectee, filler);
+    }
+
+
+    // fill bottom to top in class hierarchy as dart will also by default
+    // initialize all field before constructor bodies are run i.e. this.fields before super.
+    // see also https://www.dartlang.org/dart-tips/dart-tips-ep-11.html
+
+    var objectClassMirror = reflectClass(Object);
     ClassMirror classMirror = objMirror.type;
+    while(classMirror != null) {
+      classMirror.declarations.forEach((sym, decl) {
+        if (!decl.isPrivate &&
+            ((decl is VariableMirror && !decl.isFinal && !decl.isConst) ||
+                decl is MethodMirror)) {
+          String varName = _getName(sym);
+          String fieldName = varName;
+          TypeMirror valueType;
 
-    classMirror.declarations.forEach((sym, decl) {
-      if (!decl.isPrivate &&
-          ((decl is VariableMirror && !decl.isFinal && !decl.isConst) ||
-              decl is MethodMirror)) {
-        String varName = _getName(sym);
-        String fieldName = varName;
-        TypeMirror valueType;
+          // if it's a setter function we need to change the name
+          if (decl is MethodMirror && decl.isSetter) {
+            fieldName = varName = varName.substring(0, varName.length - 1);
+            _log.finer('Found setter function varName: ' + varName);
+            valueType = decl.parameters[0].type;
+          } else if (decl is VariableMirror) {
+            valueType = decl.type;
+          } else {
+            return;
+          }
 
-        // if it's a setter function we need to change the name
-        if (decl is MethodMirror && decl.isSetter) {
-          fieldName = varName = varName.substring(0, varName.length - 1);
-          _log.finer('Found setter function varName: ' + varName);
-          valueType = decl.parameters[0].type;
-        } else if (decl is VariableMirror) {
-          valueType = decl.type;
-        } else {
-          return;
+          // check if the property is renamed by DartsonProperty
+          Property prop = _getProperty(decl);
+          if (prop != null && prop.name != null) {
+            fieldName = prop.name;
+          }
+
+          _log.finer(
+              'Try to fill object with: ${fieldName}: ${filler[fieldName]}');
+          if (filler[fieldName] != null) {
+            objMirror.setField(new Symbol(varName),
+                _convertValue(valueType, filler[fieldName], varName, mapper));
+          }
         }
-
-        // check if the property is renamed by DartsonProperty
-        Property prop = _getProperty(decl);
-        if (prop != null && prop.name != null) {
-          fieldName = prop.name;
-        }
-
-        _log.finer(
-            'Try to fill object with: ${fieldName}: ${filler[fieldName]}');
-        if (filler[fieldName] != null) {
-          objMirror.setField(new Symbol(varName),
-              _convertValue(valueType, filler[fieldName], varName));
-        }
-      }
-    });
+      });
+      classMirror = classMirror.superclass == objectClassMirror ? null : classMirror.superclass;
+    }
 
     _log.fine("Filled object completly: ${filler}");
   }
@@ -198,7 +276,25 @@ class Dartson<T> {
   /// returns Deserialized value
   ///  Throws [IncorrectTypeTransform] if json data types doesn't match.
   ///  Throws [NoConstructorError]
-  Object _convertValue(TypeMirror valueType, Object value, String key) {
+  Object _convertValue(TypeMirror defaultValueType, Object value, String key, DecodingReferenceMapper mapper) {
+    if (mapper != null && mapper.isPlaceholder(value)) {
+      var instance = mapper.resolveReferenceForPlaceholder(value);
+      if (instance == null) {
+        _log.severe("unable to resolve reference for placeholder ${value}");
+      }
+      return instance;
+    }
+    var valueType = defaultValueType;
+
+    var identifier = value is Map ? value[TYPE_KEY] : null;
+    if (identifier != null) {
+      var type = _types[identifier];
+      if (type == null) {
+         throw new UnknownIdentifierError(identifier);
+      }
+      valueType = reflectClass(type);
+    }
+
     var symbolName = _getName(valueType.qualifiedName),
         transformer;
 
@@ -212,10 +308,10 @@ class Dartson<T> {
       _log.finer('Handle generic type.');
       // handle generic lists
       if (varMirror.originalDeclaration.qualifiedName == _QN_LIST) {
-        return _convertGenericList(varMirror, value);
+        return _convertGenericList(varMirror, value, mapper);
       } else if (varMirror.originalDeclaration.qualifiedName == _QN_MAP) {
         // handle generic maps
-        return _convertGenericMap(varMirror, value);
+        return _convertGenericMap(varMirror, value, mapper);
       }
     } else if (symbolName == 'dynamic') {
       // dynamic is used in JavaScript runtime
@@ -227,7 +323,7 @@ class Dartson<T> {
       var obj = _initiateClass(valueType);
 
       if (!(value is String) && !(value is num) && !(value is bool)) {
-        _fillObject(obj, value);
+        _fillObject(obj, value, mapper);
       } else {
         throw new IncorrectTypeTransform(value, symbolName, key);
       }
@@ -239,14 +335,14 @@ class Dartson<T> {
   }
 
   /// Converts a list of objects to a list with a Class.
-  List _convertGenericList(ClassMirror listMirror, List fillerList) {
+  List _convertGenericList(ClassMirror listMirror, List fillerList, DecodingReferenceMapper mapper) {
     _log.finer('Converting generic list');
     ClassMirror itemMirror = listMirror.typeArguments[0];
     InstanceMirror resultList = _initiateClass(listMirror);
 
     fillerList.forEach((item) {
       (resultList.reflectee as List)
-          .add(_convertValue(itemMirror, item, "@LIST_ITEM"));
+          .add(_convertValue(itemMirror, item, "@LIST_ITEM", mapper));
     });
 
     _log.finer("Created generic list: ${resultList.reflectee}");
@@ -254,7 +350,7 @@ class Dartson<T> {
   }
 
   /// Converts a generic map.
-  Map _convertGenericMap(ClassMirror mapMirror, Map fillerMap) {
+  Map _convertGenericMap(ClassMirror mapMirror, Map fillerMap, DecodingReferenceMapper mapper) {
     _log.finer('Converting generic map');
 
     ClassMirror itemMirror = mapMirror.typeArguments[1];
@@ -263,8 +359,8 @@ class Dartson<T> {
     Map reflectee = {};
 
     fillerMap.forEach((key, value) {
-      var keyItem = _convertValue(keyMirror, key, "@MAP_KEY");
-      var valueItem = _convertValue(itemMirror, value, "@MAP_VALUE");
+      var keyItem = _convertValue(keyMirror, key, "@MAP_KEY", mapper);
+      var valueItem = _convertValue(itemMirror, value, "@MAP_VALUE", mapper);
       reflectee[keyItem] = valueItem;
 
       _log.finer("Added item ${valueItem} to map key: ${keyItem}");