+++ lib/dartson_static.dart
@@ -4,9 +4,12 @@ export 'src/annotations.dart';
 export 'src/static_entity.dart';
 import 'src/static_entity.dart';
 import './type_transformer.dart';
+import 'src/reference_mapper.dart';
+import 'src/shared_exceptions.dart';
 
 import 'package:logging/logging.dart';
 import 'dart:convert' show Codec, JSON;
+import 'dart:mirrors';
 
 final _SimpleTypeTransformer _st = new _SimpleTypeTransformer();
 final Map<Type, TypeTransformer> _transformers = {
@@ -26,11 +29,76 @@ class _SimpleTypeTransformer extends TypeTransformer {
   encode(value) => value;
 }
 
+class _StaticEntityEncoderImpl extends StaticEntityEncoder {
+  final Map<Type, TypeTransformer> transformers;
+  final Map<String, Type> types;
+  final EncodingReferenceMapper mapper;
+
+  _StaticEntityEncoderImpl(this.transformers, this.types, this.mapper);
+
+  bool hasTransformer(Type type) => transformers[type] != null;
+  TypeTransformer getTransformer(Type type) => transformers[type];
+
+  bool isSerialized(Object instance) {
+    return mapper != null && mapper.isSerialized(instance);
+  }
+   Map createSerializablePlaceholder(Object instance) {
+     return mapper.createSerializablePlaceholder(instance);
+   }
+   void registerSerializableMap(Object instance, dynamic serializedObject) {
+     if (mapper != null) {
+       mapper.registerSerializableMap(instance, serializedObject);
+     }
+     var key = types.keys.firstWhere((key) => types[key] == instance.runtimeType, orElse: () => null);
+     if (key != null) {
+       serializedObject.putIfAbsent(TYPE_KEY, () => key);
+     }
+   }
+}
+
+class _StaticEntityDecoderImpl extends StaticEntityDecoder {
+  final Map<Type, TypeTransformer> transformers;
+  final Map<String, Type> types;
+  final DecodingReferenceMapper mapper;
+
+  _StaticEntityDecoderImpl(this.transformers, this.types, this.mapper);
+
+
+  Object createInstance(Map serializableMap, Object defaultInstance()) {
+    var identifier = serializableMap[TYPE_KEY];
+    if (identifier == null) {
+      return defaultInstance();
+    }
+    var type = types[identifier];
+    if (type == null) {
+      throw new UnknownIdentifierError(identifier);
+    }
+    var c = reflectClass(type);
+    return c.newInstance(new Symbol(''), []).reflectee;
+  }
+
+  bool hasTransformer(Type type) => transformers[type] != null;
+  TypeTransformer getTransformer(Type type) => transformers[type];
+
+  void registerInstanceIfApplicable(Object instance, Map serializableMap) {
+    if (mapper != null) {
+      mapper.registerInstanceIfApplicable(instance, serializableMap);
+    }
+  }
+  bool isPlaceholder(Object val) {
+    return mapper != null && mapper.isPlaceholder(val);
+  }
+  Object resolveReferenceForPlaceholder(Map placeholder) {
+    return mapper.resolveReferenceForPlaceholder(placeholder);
+  }
+}
+
 /// Static version of dartson.
-class Dartson extends TypeTransformerProvider {
+class Dartson  {
   final Codec _codec;
   final Logger _log;
   final Map<Type, TypeTransformer> transformers = {};
+  final Map<String, Type> _types = {};
 
   Dartson(this._codec, [String identifier = 'dartson'])
       : _log = new Logger(identifier) {
@@ -45,10 +113,26 @@ class Dartson extends TypeTransformerProvider {
     transformers[type] = transformer;
   }
 
-  bool hasTransformer(Type type) => transformers[type] != null;
-  TypeTransformer getTransformer(Type type) => transformers[type];
+  addIdentifier(String identifier, Type type) {
+    _types[identifier] = type;
+  }
 
-  Object map(Object data, StaticEntity clazz, [bool isList = false]) {
+  Object map(Object data, StaticEntity entity, [bool isList = false, bool beReferenceAware = false]) {
+    var clazz = entity;
+      if (clazz == null) {
+        var identifier = data is Map ? data[TYPE_KEY] : null;
+        if (identifier == null) {
+          throw new NullObjectError();
+        }
+        var type = _types[identifier];
+        if (type == null) {
+          throw new UnknownIdentifierError(identifier);
+        }
+        var c = reflectClass(type);
+        clazz = c.newInstance(new Symbol(''), []).reflectee;
+      }
+
+    var staticEntityDecoder = new _StaticEntityDecoderImpl(this.transformers, this._types, beReferenceAware ? new DecodingReferenceMapper() : null);
     if (data is List && isList) {
       List returnList = [];
 
@@ -63,7 +147,7 @@ class Dartson extends TypeTransformerProvider {
           cl = clazz.newEntity();
         }
 
-        cl.dartsonEntityDecode(item, this);
+        cl.dartsonEntityDecode(item, staticEntityDecoder);
         returnList.add(cl);
       });
 
@@ -71,20 +155,20 @@ class Dartson extends TypeTransformerProvider {
     } else if (data is List || isList) {
       throw 'Incompatible none list type to list.';
     } else {
-      clazz.dartsonEntityDecode(data, this);
+      clazz.dartsonEntityDecode(data, staticEntityDecoder);
       return clazz;
     }
   }
 
-  Object serialize(Object data, {Type type}) {
+  Object serialize(Object data, {Type type, EncodingReferenceMapper mapper}) {
     var transformer;
 
     if (data is List) {
-      return _serializeList(data);
+      return _serializeList(data, mapper);
     } else if (data is Map) {
-      return _serializeMap(data);
+      return _serializeMap(data, mapper);
     } else if (data is StaticEntity) {
-      return data.dartsonEntityEncode(this);
+      return data.dartsonEntityEncode(new _StaticEntityEncoderImpl(this.transformers, this._types, mapper));
     } else if (type != null && (transformer = transformers[type]) != null) {
       return transformer.encode(data);
     } else {
@@ -92,22 +176,30 @@ class Dartson extends TypeTransformerProvider {
     }
   }
 
-  dynamic encode(Object clazz) {
-    return _codec.encode(serialize(clazz));
+  dynamic encodeReferenceAware(Object clazz) {
+    return encode(clazz, beReferenceAware: true);
   }
 
-  dynamic decode(var encoded, Object object, [bool isList = false]) {
-    return map(_codec.decode(encoded), object, isList);
+  dynamic encode(Object clazz, {bool beReferenceAware: false}) {
+    return _codec.encode(serialize(clazz, mapper: beReferenceAware ? new EncodingReferenceMapper() : null));
   }
 
-  List _serializeList(List list) {
-    return list.map((i) => serialize(i)).toList();
+  dynamic decodeReferenceAware(var encoded, Object object, [ bool isList = false]) {
+    return decode(encoded, object, isList, true);
   }
 
-  Map _serializeMap(Map map) {
+  dynamic decode(var encoded, Object object, [bool isList = false, bool beReferenceAware =  false]) {
+    return map(_codec.decode(encoded), object, isList, beReferenceAware);
+  }
+
+  List _serializeList(List list, [EncodingReferenceMapper mapper]) {
+    return list.map((i) => serialize(i, mapper: mapper)).toList();
+  }
+
+  Map _serializeMap(Map map, [EncodingReferenceMapper mapper]) {
     Map newMap = new Map<String, Object>();
     map.forEach((key, val) {
-      if (val != null) newMap[key] = serialize(val);
+      if (val != null) newMap[key] = serialize(val, mapper: mapper);
     });
 
     return newMap;