+++ test/transformer_test.dart
@@ -1,6 +1,8 @@
 library dartson.transformer.test;
 
 import 'dart:io';
+import 'dart:isolate';
+import 'dart:async';
 
 import '../lib/transformer.dart';
 
@@ -11,6 +13,8 @@ import 'package:test/test.dart';
 String get _testDirPath => p.dirname(p.fromUri(Platform.script));
 
 void main() {
+  var tempDir;
+
   // test the compiler
   var simplePath = p.join(_testDirPath, 'fixture/simple_class.dart');
   FileCompiler compiler = new FileCompiler(simplePath);
@@ -19,6 +23,12 @@ void main() {
   });
 
   CompilationUnitMember simpleClass;
+
+  tearDown(() {
+    if (tempDir != null) {
+      tempDir.deleteSync(recursive: true);
+    }
+  });
 
   test('should contain the SimpleClass declaration', () {
     expect(compiler.compilationUnit.declarations.any((CompilationUnitMember m) {
@@ -43,7 +53,10 @@ void main() {
 
   test('the compiler should contain the SimpleClass', () {
     expect(compiler.entities.any((m) => m.name.name == 'SimpleClass'), true);
+
+
   });
+
 
   List<PropertyDefinition> entityMap;
   test('fetch entity map from simpleClass', () {
@@ -74,4 +87,71 @@ void main() {
     var newFile = new File(p.join(tempDir.path, 'part1_class.dart'));
     newFile.writeAsStringSync(newCode);
   });
+
+  test('build, compile and run code for circular referenced model', () {
+    tempDir = Directory.systemTemp.createTempSync('dartson_');
+
+    var compiledFixture = compileFixtureToDirectory('circular_referenced_model.dart', tempDir);
+    var fixtureTest = generateFixtureTest(
+        compiledFixture,
+        tempDir,
+        'reference_aware_test.dart',
+        'testSerializeAndDeserializeReferenceAware');
+
+    runTestIsolated(fixtureTest);
+
+  });
+
+  test('build, compile and run code for polymorphic model', () {
+    tempDir = Directory.systemTemp.createTempSync('dartson_');
+
+    var compiledFixture = compileFixtureToDirectory('polymorphic_model.dart', tempDir);
+    var fixtureTest = generateFixtureTest(
+        compiledFixture,
+        tempDir,
+        'polymorphic_test.dart',
+        'testSerializeAndDeserializePolymorphic');
+
+    runTestIsolated(fixtureTest);
+  });
+}
+
+File compileFixtureToDirectory(String dartFile, Directory toDir) {
+  var newCompiler = new FileCompiler(p.join(_testDirPath, 'fixture/$dartFile'));
+  var newCode = newCompiler.build('package:dartson/test/$dartFile');
+  var newFile = new File(p.join(toDir.path, 'fixture/$dartFile'))..createSync(recursive: true);
+  newFile.writeAsStringSync(newCode);
+  return newFile;
+}
+
+File generateFixtureTest(File compiledFixture, Directory toDir, String sharedTestHelper, String testMethod) {
+  var targetSharedTestHelper = new File(p.join(toDir.path, 'shared/${sharedTestHelper}'))
+    ..createSync(recursive: true);
+  new File(p.join(_testDirPath, 'shared/${sharedTestHelper}'))
+    ..copySync(targetSharedTestHelper.path);
+
+  var testRunnerFile = new File(p.join(toDir.path, 'testrunner.dart'));
+  testRunnerFile.writeAsStringSync('''
+  library dartson_test.test;
+
+  import 'package:dartson/dartson_static.dart' as ds;
+  import 'package:test/test.dart';
+  import 'dart:isolate';
+
+  import './shared/${sharedTestHelper}';
+
+  void main(List<String> args, SendPort replyTo) {
+    $testMethod(() => new ds.Dartson.JSON()).then((success) => replyTo.send(success));
+  }
+
+  ''');
+  return testRunnerFile;
+}
+
+void runTestIsolated(fixtureTest) {
+  var done = expectAsync((){});
+  var response = new ReceivePort();
+  Isolate.spawnUri(Uri.parse(fixtureTest.path), [], response.sendPort)
+    .then((_) => response.first)
+    .then((success) { expect(success, true); done(); });
 }